split seed 49 page component
