+++ src/components/card/TitleCard.tsx
@@ -3,21 +3,16 @@ import { useMemo } from 'react';
 
 interface TitleCardProps {
     title: string;
-    marginRight?: number;
 }
 
 const TitleCard = (props: TitleCardProps) => {
-
     return useMemo(() => (
         <Card elevation={0}
               variant={'outlined'}
               sx={(theme) => ({
                   marginBottom: theme.spacing(1),
-                  marginRight: props.marginRight
-                      ? theme.spacing(props.marginRight)
-                      : undefined,
               })}>
-            <CardContent sx={{ display: 'flex' }}>
+            <CardContent>
                 <Typography variant={'h1'}>
                     {props.title}
                 </Typography>
