+++ src/components/display/Display.tsx
@@ -1,0 +1,13 @@
+import { PropsWithChildren } from 'react';
+
+interface DisplayProps {
+    when: boolean;
+}
+
+const Display = ({ when, children }: PropsWithChildren<DisplayProps>) => {
+    return when
+        ? <>{children}</>
+        : null;
+};
+
+export default Display;
