+++ src/components/display/index.ts
@@ -1,0 +1,1 @@
+export { default as Display } from './Display';
